52 improve coverage two small things to fix (#56)
* improve cylinder tests

* imporve point at function

* delete temp mesh files

* remove comment

* rename cylinder test file

* [compression cylinder] use tempfile and gmshio instead of meshio

* attempt to make style check pass

---------
